hotfix: Make sure things that should be filtered are filtered
This is a hack, and should be reverted once we get to the bottom of it.
It will impact performance, but right now there are things that should
be filtered and are not, so lets get this merged for security purposes.
Hopefully this library will be re-written soon.

Change-type: patch
See: https://github.com/balena-io/jellyfish/pull/878
Signed-off-by: Juan Cruz Viotti <juan@balena.io>